# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Registro Epidemiologico de Doenças Raras",
        "description": "Aplicação REST que consiste em um registro epidemiológico para pacientes com doenças raras a\n    servir aplicações que podem ser outros sistemas ou uma interface de usuário.",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "localhost:8000/registro_epidemiologico/api",
            "description": ""
        },
        {
            "url": "localhost:8000/registro_epidemiologico/"
        }
    ],
    "paths": {
...
        "/doenca": {
            "get": {
                "tags": [
                    "doença"
                ],
                "summary": "Lista de doenças",
                "description": "Retorna uma lista com as doenças cadastradas.",
                "operationId": "index",
                "responses": {
                    "401": {
                        "description": "Não autorizado"
                    },
                    "200": {
                        "description": "Operação realizada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": "1"
                                        },
                                        "nome": {
                                            "type": "string",
                                            "example": "Doença de Sandhoff"
                                        },
                                        "descricao": {
                                            "type": "string",
                                            "example": "São esfingolipidoses, distúrbios hereditários do metabolismo, causadas por deficiência de hexosaminidase que levam a sintomas neurológicos graves e morte prematura."
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "doença"
                ],
                "summary": "Cadastrar doença",
                "description": "Cadastra doença. Retorna a doença cadastrada.",
                "operationId": "store",
                "responses": {
                    "201": {
                        "description": "Registro cadastrado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": "1"
                                        },
                                        "nome": {
                                            "type": "string",
                                            "example": "Doença de Sandhoff"
                                        },
                                        "descricao": {
                                            "type": "string",
                                            "example": "São esfingolipidoses, distúrbios hereditários do metabolismo, causadas por deficiência de hexosaminidase que levam a sintomas neurológicos graves e morte prematura."
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/doenca/{doenca_id}": {
            "get": {
                "tags": [
                    "doença"
                ],
                "summary": "Listar doença",
                "description": "Retorna dados da doença cujo id é passado pela url.",
                "operationId": "show",
                "parameters": [
                    {
                        "name": "doenca_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "int"
                        }
                    }
                ],
                "responses": {
                    "401": {
                        "description": "Não autorizado"
                    },
                    "404": {
                        "description": "Registro não encontrado"
                    },
                    "200": {
                        "description": "Operação realizada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer"
                                        },
                                        "nome": {
                                            "type": "string"
                                        },
                                        "descricao": {
                                            "type": "string"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "doença"
                ],
                "summary": "Atualizar doença",
                "description": "Atualiza dados da doença cujo id é passado na url. Retorna a doença com os dados atualizados.",
                "operationId": "update",
                "parameters": [
                    {
                        "name": "doenca_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "int"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Atualizar doença já cadastrada",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nome",
                                    "descricao"
                                ],
                                "properties": {
                                    "nome": {
                                        "type": "string",
                                        "format": "",
                                        "example": "Doença de Sandhoff"
                                    },
                                    "descricao": {
                                        "type": "string",
                                        "format": "",
                                        "example": "São esfingolipidoses, distúrbios hereditários do metabolismo, causadas por deficiência de hexosaminidase que levam a sintomas neurológicos graves e morte prematura."
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Operação realizada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": "1"
                                        },
                                        "nome": {
                                            "type": "string",
                                            "example": "Doença de Sandhoff"
                                        },
                                        "descricao": {
                                            "type": "string",
                                            "example": "São esfingolipidoses, distúrbios hereditários do metabolismo, causadas por deficiência de hexosaminidase que levam a sintomas neurológicos graves e morte prematura."
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autorizado"
                    },
                    "400": {
                        "description": "Registro já cadastrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "doença"
                ],
                "summary": "Remover doença",
                "description": "Remove a doença cujo id é passado pela url.",
                "operationId": "destroy",
                "parameters": [
                    {
                        "name": "doenca_id",
                        "in": "path",
                        "description": "id da doença a ser removida",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "401": {
                        "description": "Não autorizado"
                    },
                    "404": {
                        "description": "Redistro não encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/uf/{if_id}": {
            "put": {
                "tags": [
                    "unidade federativa"
                ],
                "summary": "Atualiza a UF cujo id é passado pela url.",
                "description": "Atualiza a UF. Retorna a UF atualizada.",
                "operationId": "update",
                "parameters": [
                    {
                        "name": "uf_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "int"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operação realizada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": "1"
                                        },
                                        "nome": {
                                            "type": "string",
                                            "example": "Paraíba"
                                        },
                                        "sigla": {
                                            "type": "string",
                                            "example": "PB"
                                        },
                                        "codigo_ibge": {
                                            "type": "integer",
                                            "example": "25"
                                        },
                                        "regia_id": {
                                            "type": "integer",
                                            "example": "PB"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "unidade federativa"
                ],
                "summary": "Deleta UF",
                "description": "Remove a UF cujo id é passado pela url.",
                "operationId": "destroy",
                "parameters": [
                    {
                        "name": "uf_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "int"
                        }
                    }
                ],
                "responses": {
                    "401": {
                        "description": "Não autorizado"
                    },
                    "404": {
                        "description": "Registro não encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "apiKey",
                "description": "Apenas usuários autenticados podem realizar esta operação",
                "name": "token",
                "in": "header"
            }
        }
    },
    "security": [
        []
    ]
}